requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
neo4j==5.14.0
neo4j-driver==5.27.0
openai==1.59.3
streamlit==1.28.2
pillow==10.1.0
pytesseract==0.3.10
numpy==1.26.2
pydantic==2.5.1
pydantic-settings==2.7.1
networkx 
plotly 
pytest 
pytest-asyncio